requirements.txt
# requirements.txt
crewai>=0.28.8
litellm>=1.34.2
playwright>=1.40.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pandas>=2.1.0
openai>=1.3.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.1.0
streamlit>=1.28.0
schedule>=1.2.0